+++ NEWS.md
@@ -5,6 +5,16 @@ All notable changes to ´Manifolds.jl´ will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.10] 2025-12-xx
+
+### Added
+
+* Chart-based variants of `christoffel_symbols_first`, `einstein_tensor`, `det_local_metric` and `log_local_metric_density`.
+
+### Changed
+
+* Deprecated variants of `det_local_metric`, `einstein_tensor`, `inverse_local_metric`, `local_metric`, `local_metric_jacobian`, `ricci_curvature` that receive a point and a basis as arguments. New variants that work in charts should be used instead.
+
 ## [0.11.9] 2025-12-15
 
 ### Fixed